requirements.txt
APScheduler==3.6.3
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
Django==3.2.2
django-cors-headers==3.7.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
idna==2.10
parameterized==0.8.1
Pillow==8.2.0
PyJWT==2.1.0
python-dotenv==0.17.1
python-telegram-bot==13.5
pytz==2021.1
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
tornado==6.1
tzlocal==2.1
urllib3==1.26.5
